src/ml-metrics/requirements.txt
prometheus-client==0.16.0
psycopg2-binary==2.9.6